obstruction - refactor
Make the player making the move a field and initialise in the constructor.
This way it can be final and cannot change.
The Round class should be a fixed state in time.
The 2D array breaks that but at least we can do that for the player.
We should know the players and their order when creating the Round(s).
The orchestration of the game should be left to someone else.
They already do it since they pass that knowledge in the play method.

*Considerations*
As it was before one could probably use one Round for the whole game
alternating the player every play.
Like this, the player is fixed for the round which semantically makes
sense BUT for now, one can make an arbitrary number of moves and all
would be locked to the same player...
So...
We could make the play method single use only...
I am wondering if maybe that means that the play method should be called
from the constructor as well.
So constructing new round takes the old board, the new player and their move.